traits example
